configure phoenix apps
